requirements.txt
streamlit>=1.24.0
python-dotenv==0.19.0
openai==0.27.0
loguru==0.6.0
sounddevice==0.4.6
soundfile==0.12.1
numpy>=1.20.0